permissions for deleting event

+++ events/views.py
@@ -18,6 +18,7 @@ from models import EventModel, HikeEventModel, DinnerEventModel, GenericTripMode
 from invite.models import InviteModel, MembershipModel
 from users.forms import UserRegistrationForm
 from base.helpers import getMenuInfo
+from base.permissions import memberCheck, isCreator
 
 
 @login_required(login_url = '/loginRequired/')
@@ -44,7 +45,13 @@ def editMembers(request, eventid):
 
 @login_required(login_url = '/loginRequired/')
 def deleteEvent(request, eventid):
+	
 	event = EventModel.objects.filter(eventid=eventid)
+	if memberCheck(request.user, event[0]) == False:
+		return render(request, 'invite/notMember.html', {'menu' : getMenuInfo(request), 'title' : 'Not Member'})
+	if isCreator(request.user, event[0]) == False:
+		return render(request, 'events/notPermission.html', {'menu' : getMenuInfo(request), 'title' : 'Not Permission'})
+	
 	eventChild = EventModel.getEvent(eventid)
 	eventChild.delete()
 	event.delete()
